Move CurrencySymbol modifier to its own file
